feat: create a new release #47
This commit adds new bb tooling to create a new release. This includes
setting the new version of the release in the repo.edn file and
creating and actual release using the GH API.

Also updates the gh action to automatically create a commit for that
release when changes are merged into trunk.

epic: #9