refactor: change how tooling is used #101
This commit changes how the tooling prodived by clci works and how it is
used. Most of the tooling was changed to be invoked using the mechanism
and ad-hoc Actions.

The tools are no longer invoked by babashka tasks, instead there is a
central clci task with its own dispatch method to run tools.