feat: automatically generate a changelog #94
This commit resolves #94 to automatically create and update the
changelog. It provides an Action and an adhoc job to update the
changelog.

The changelog is generated from the git commit log by taking all commits
since the latest release which is retrieved from the scm provider. It
can either create a new release entry or add the changes to the
unreleased section of the changelog.

The changelog itself loosely follows the specification from
keepachangelog.com.