feat: add optional trunk branch arg #143
This commit adds an extra (optional) argument to the release task to
specify a branch name to use as trunk. This allows to use a branch
naming scheme where the trunk is named different than _master_.

This commit also allows to set the trunk branch in the _repo.edn_ file.