fix: fixes #153
This commit resolves the bug that no changelog updates were possible
when trunk is not master. This fix allows to either give an explicit
command line argument to specify the trunk branch when running the
release task or to use the trunk as configured in the repo.edn file. as
a fallback master is used as trunk.

This commit also updates the ci pipeline and removes alternative cache
keys.